+++ src/tabs/inflection.ts
@@ -3,44 +3,66 @@ import Expressions from "../constants/expressions.js";
 import Selectors from "../constants/selectors.js";
 import { ContentTabs, SearchOptionsWithWord } from "../options.js";
 
-export interface InflectionTable extends Header, Body {}
+/** Represents a table of inflections as parsed directly from the tables on Dexonline. */
+export interface InflectionModel extends Heading, Table {}
 
-interface Header {
-	tags: string[];
-	index: number;
-	lemma: string;
+/** Represents the heading of a Dexonline inflection model. */
+interface Heading {
+  /** The tags for a given inflection model. */
+	readonly tags: string[];
+  /** The index of the inflection model as shown on the webpage. */
+	readonly index: number;
+  /** The lemma this inflection model applies to */
+	readonly lemma: string;
 }
 
-interface Body {
-	table: string[][];
+/** Represents the body (table) of a Dexonline inflection model. */
+interface Table {
+  /** The HTML table represented as a 2D array of rows and columns. */
+	readonly table: string[][];
 }
 
-export function parse($: CheerioAPI, options: SearchOptionsWithWord): InflectionTable[] {
+/**
+ * Given a {@link $|Cheerio document handle} and additional {@link options} for scraping entries, scrapes the
+ * inflection models on the page.
+ * 
+ * @param $ - A Cheerio document handle for the webpage.
+ * @param options - Options for the scraper.
+ * @returns An array of the scraped {@link InflectionModel}s.
+ */
+export function scrape($: CheerioAPI, options: SearchOptionsWithWord): InflectionModel[] {
 	const inflection = $(Selectors.contentTab(ContentTabs.Inflection));
 
 	const entries = inflection.find(Selectors.contentTabs.inflection.entry.element).toArray();
 
-	const tables: InflectionTable[] = [];
+	const tables: InflectionModel[] = [];
 	for (const entry of entries) {
 		const tableElement = $(entry).children(Selectors.contentTabs.inflection.entry.table.element).first();
 
-		const header = parseHeader($, tableElement);
-		if (options.mode === "strict" && header.lemma !== options.word) {
+		const heading = scrapeHeading($, tableElement);
+		if (options.mode === "strict" && heading.lemma !== options.word) {
 			continue;
 		}
 
-		const body = parseBody($, tableElement);
+		const body = scrapeTable($, tableElement);
 		if (body.table.length === 0) {
 			continue;
 		}
 
-		tables.push({ ...header, ...body });
+		tables.push({ ...heading, ...body });
 	}
 	return tables;
 }
 
-function parseHeader($: CheerioAPI, header: Cheerio<Element>): Header {
-	const section = header.children(Selectors.contentTabs.inflection.entry.table.header.element);
+/**
+ * Given a {@link $|Cheerio document handle} for the inflection model on the webpage, scrapes its heading.
+ * 
+ * @param $ - A Cheerio document handle for the webpage.
+ * @param $heading - A Cheerio document handle for the heading of the inflection model.
+ * @returns The scraped inflection model {@link Heading}.
+ */
+function scrapeHeading($: CheerioAPI, $heading: Cheerio<Element>): Heading {
+	const section = $heading.children(Selectors.contentTabs.inflection.entry.table.header.element);
 
 	const lemmaString = section.children(Selectors.contentTabs.inflection.entry.table.header.lemma).html() ?? undefined;
 	if (lemmaString === undefined) {
@@ -69,7 +91,14 @@ function parseHeader($: CheerioAPI, header: Cheerio<Element>): Header {
 	return { tags, lemma, index };
 }
 
-function parseBody($: CheerioAPI, body: Cheerio<Element>): Body {
+/**
+ * Given a {@link $|Cheerio document handle} for the inflection model on the webpage, scrapes its table.
+ * 
+ * @param $ - A Cheerio document handle for the webpage.
+ * @param $heading - A Cheerio document handle for the table of the inflection model.
+ * @returns The scraped inflection model {@link Table}.
+ */
+function scrapeTable($: CheerioAPI, body: Cheerio<Element>): Table {
 	const section = body.children(Selectors.contentTabs.inflection.entry.table.body.element);
 
 	// Certain words are listed in the inflection tab but do not show up with a table.
@@ -130,7 +159,7 @@ function parseBody($: CheerioAPI, body: Cheerio<Element>): Body {
 			const rowSpan = Number(column.attr("rowspan") ?? "1");
 			const columnSpan = Number(column.attr("colspan") ?? "1");
 
-			const text = getCellText($, column);
+			const text = scrapeCellContents($, column);
 
 			for (const _ of Array(columnSpan).keys()) {
 				const index = freeCells.shift() ?? extendedRows.length;
@@ -151,14 +180,22 @@ function parseBody($: CheerioAPI, body: Cheerio<Element>): Body {
 	return { table };
 }
 
-function getCellText($: CheerioAPI, cell: Cheerio<Element>): string {
-	const listElements = cell
+/**
+ * Given a {@link $|Cheerio document handle} for a cell inside the table for an inflection model, gets the
+ * contents of the cell.
+ * 
+ * @param $ - A Cheerio document handle for the webpage.
+ * @param $heading - A Cheerio document handle for the cell element.
+ * @returns The scraped contents of the cell.
+ */
+function scrapeCellContents($: CheerioAPI, $cell: Cheerio<Element>): string {
+	const listElements = $cell
 		.children("ul")
 		.children("li")
 		.map((_, element) => $(element));
 
 	if (listElements.length === 0) {
-		return cell.text().trim().replaceAll(/ +/g, " ");
+		return $cell.text().trim().replaceAll(/ +/g, " ");
 	}
 
 	const parts: string[] = [];